Fix Geoffrey, SirKay, complete diary command, start hand in the sand quest

+++ Server/src/main/core/game/node/entity/player/link/diary/Diary.java
@@ -182,7 +182,7 @@ public class Diary {
             }
             if (complete) {
                 completedLevels.add(level);
-            } else if (completedLevels.contains(level)) completedLevels.remove(level);
+            } else if (completedLevels.contains(level)) completedLevels.remove((Integer) level);
         }
     }
